:construction:  Methods for construct builder menus in console apps
[ADD] Builders for menus in console

Added ConsoleMenuBuilder and ConsoleMenuReadBuilder classes to construct menus in console apps.
The builders allow the user to select an option from a list of options, and execute an action based on the selection.
[ADD] Interface to extract some functions of the console, to do unit tests

Added IUserInputReader interface to abstract console input and allow for unit testing of the builders.
[ADD] Unit tests

Added unit tests to ensure the proper functioning of the ConsoleMenuBuilder and ConsoleMenuReadBuilder classes.

[ADD] Method to read integer input in ConsoleMenuReadBuilder

Added a new method to ConsoleMenuReadBuilder that allows the user to read an integer input from the console.